F_T30603 move MapBundle logic to core
- adjust use statements
