+++ demosplan/DemosPlanCoreBundle/Logic/Maps/MapCoordinateDataFetcher.php
@@ -14,9 +14,9 @@ namespace demosplan\DemosPlanCoreBundle\Logic\Maps;
 
 use demosplan\DemosPlanCoreBundle\Exception\ExternalDataFetchException;
 use demosplan\DemosPlanCoreBundle\Logic\LocationService;
+use demosplan\DemosPlanCoreBundle\Logic\Map\MapService;
 use demosplan\DemosPlanCoreBundle\ValueObject\LocationData;
 use demosplan\DemosPlanCoreBundle\ValueObject\MapCoordinate;
-use demosplan\DemosPlanMapBundle\Logic\MapService;
 use demosplan\DemosPlanStatementBundle\Exception\InvalidDataException;
 use Geocoder\Provider\Provider;
 use Geocoder\Query\ReverseQuery;
