+++ demosplan/DemosPlanMapBundle/Utilities/MapScreenshotter.php
@@ -15,18 +15,18 @@ namespace demosplan\DemosPlanMapBundle\Utilities;
 use DemosEurope\DemosplanAddon\Contracts\ApiClientInterface;
 use DemosEurope\DemosplanAddon\Contracts\Config\GlobalConfigInterface;
 use DemosEurope\DemosplanAddon\Utilities\Json;
+use demosplan\DemosPlanCoreBundle\Logic\Map\FeaturesToMapLayersConverter;
+use demosplan\DemosPlanCoreBundle\Logic\Map\GeoJsonToFeaturesConverter;
+use demosplan\DemosPlanCoreBundle\Logic\Map\MapImageToCoordinatesCropper;
+use demosplan\DemosPlanCoreBundle\Logic\Map\MapImageToPolygonCropper;
+use demosplan\DemosPlanCoreBundle\Logic\Map\MapLayerMerger;
+use demosplan\DemosPlanCoreBundle\Logic\Map\MinCoordinatesExtractor;
+use demosplan\DemosPlanCoreBundle\Logic\Map\PolygonIntoMapLayerMerger;
+use demosplan\DemosPlanCoreBundle\Logic\Map\WmsToWmtsCoordinatesConverter;
 use demosplan\DemosPlanCoreBundle\Logic\Maps\WktToGeoJsonConverter;
 use demosplan\DemosPlanCoreBundle\Logic\TextIntoImageInserter;
 use demosplan\DemosPlanCoreBundle\Logic\UrlFileReader;
 use demosplan\DemosPlanCoreBundle\Utilities\DemosPlanPath;
-use demosplan\DemosPlanMapBundle\Logic\FeaturesToMapLayersConverter;
-use demosplan\DemosPlanMapBundle\Logic\GeoJsonToFeaturesConverter;
-use demosplan\DemosPlanMapBundle\Logic\MapImageToCoordinatesCropper;
-use demosplan\DemosPlanMapBundle\Logic\MapImageToPolygonCropper;
-use demosplan\DemosPlanMapBundle\Logic\MapLayerMerger;
-use demosplan\DemosPlanMapBundle\Logic\MinCoordinatesExtractor;
-use demosplan\DemosPlanMapBundle\Logic\PolygonIntoMapLayerMerger;
-use demosplan\DemosPlanMapBundle\Logic\WmsToWmtsCoordinatesConverter;
 use demosplan\DemosPlanMapBundle\ValueObject\CoordinatesViewport;
 use demosplan\DemosPlanMapBundle\ValueObject\MapLayer;
 use Exception;
