+++ tests/backend/core/Map/Functional/MapScreenshotterTest.php
@@ -11,8 +11,8 @@
 namespace Tests\Core\Map\Functional;
 
 use DemosEurope\DemosplanAddon\Utilities\Json;
-use demosplan\DemosPlanMapBundle\Logic\GeoJsonToFeaturesConverter;
-use demosplan\DemosPlanMapBundle\Logic\PolygonIntoMapLayerMerger;
+use demosplan\DemosPlanCoreBundle\Logic\Map\GeoJsonToFeaturesConverter;
+use demosplan\DemosPlanCoreBundle\Logic\Map\PolygonIntoMapLayerMerger;
 use demosplan\DemosPlanMapBundle\Utilities\MapScreenshotter;
 use Tests\Base\FunctionalTestCase;
