+++ tests/backend/core/Map/Unit/FeaturesToMapLayersConverterTest.php
@@ -12,21 +12,19 @@ declare(strict_types=1);
 
 namespace Tests\Core\Map\Unit;
 
+use demosplan\DemosPlanCoreBundle\Logic\Map\FeaturesToMapLayersConverter;
 use demosplan\DemosPlanCoreBundle\Utilities\DemosPlanPath;
-use demosplan\DemosPlanMapBundle\Logic\FeaturesToMapLayersConverter;
 use demosplan\DemosPlanMapBundle\ValueObject\CoordinatesViewport;
 use demosplan\DemosPlanMapBundle\ValueObject\Feature;
 use demosplan\DemosPlanMapBundle\ValueObject\MapLayer;
 use demosplan\DemosPlanMapBundle\ValueObject\PrintLayer;
 use demosplan\DemosPlanMapBundle\ValueObject\PrintLayerTile;
 use demosplan\DemosPlanMapBundle\ValueObject\PrintLayerTilePosition;
-
-use function imagecolorallocate;
-
 use Intervention\Image\ImageManager;
 use Point;
 use Tests\Base\UnitTestCase;
 use Tightenco\Collect\Support\Collection;
+use function imagecolorallocate;
 
 class FeaturesToMapLayersConverterTest extends UnitTestCase
 {
