+++ tests/backend/core/Map/Unit/PrintLayerToMapLayerConverterTest.php
@@ -12,18 +12,16 @@ declare(strict_types=1);
 
 namespace Tests\Core\Map\Unit;
 
+use demosplan\DemosPlanCoreBundle\Logic\Map\PrintLayerToMapLayerConverter;
 use demosplan\DemosPlanCoreBundle\Utilities\DemosPlanPath;
-use demosplan\DemosPlanMapBundle\Logic\PrintLayerToMapLayerConverter;
 use demosplan\DemosPlanMapBundle\ValueObject\CoordinatesViewport;
 use demosplan\DemosPlanMapBundle\ValueObject\PrintLayer;
 use demosplan\DemosPlanMapBundle\ValueObject\PrintLayerTile;
 use demosplan\DemosPlanMapBundle\ValueObject\PrintLayerTilePosition;
-
-use function imagecolorallocate;
-
 use Intervention\Image\ImageManager;
 use Tests\Base\UnitTestCase;
 use Tightenco\Collect\Support\Collection;
+use function imagecolorallocate;
 
 class PrintLayerToMapLayerConverterTest extends UnitTestCase
 {
